Added a vert_norm_color_tangent class and fixed a compile warning in collision_detect. -FG

+++ src/collision_detect.cpp
@@ -804,7 +804,7 @@ void coll_obj_group::set_coll_obj_props(int index, int type, float radius, float
 void coll_obj_group::remove_index_from_ids(int index) {
 
 	if (index < 0) return;
-	assert(index < size());
+	assert((size_t)index < size());
 	if (operator[](index).status == COLL_DYNAMIC) coll_objects.dynamic_ids.must_erase (index);
 	if (operator[](index).cp.draw               ) coll_objects.drawn_ids.must_erase   (index);
 	if (operator[](index).platform_id >= 0      ) coll_objects.platform_ids.must_erase(index);
